refactor: consolidate duplicate hooks into shared types and helpers
Extract 6 identical hook interfaces (ComparativeHooks, AsyncComparativeHooks,
ThroughputHooks, AsyncThroughputHooks, QuantileHooks, AsyncQuantileHooks) and
their shared execution helpers (measureSync, measureAsync, warmupSync,
warmupAsync, runSyncWithHooks, runAsyncWithHooks) into a single hooks.ts
module, eliminating cross-file duplication across the three matcher files.

No behavioral changes — all existing tests pass unchanged.

Ref: #85
