+++ src/matchers-comparative.ts
@@ -1,48 +1,12 @@
-import {nowInMillis} from "./timing";
 import {validateCallback, validateComparativeOptions} from "./validators";
 import {processComparativeResults} from "./helpers";
+import {SyncHooks, AsyncHooks, SyncCallback, AsyncCallback, runSyncWithHooks, runAsyncWithHooks, measureSync, measureAsync} from "./hooks";
 
-interface ComparativeHooks {
-  setup?: () => unknown;
-  teardown?: (suiteState: unknown) => void;
-  setupEach?: (suiteState: unknown) => unknown;
-  teardownEach?: (suiteState: unknown, iterState: unknown) => void;
-}
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any -- expect.extend erases generics at runtime
-type SyncCallback = (...args: any[]) => unknown;
-
-function warmupSync(callbackA: SyncCallback, callbackB: SyncCallback, warmupCount: number, suiteState: unknown, hooks: ComparativeHooks): void {
+function warmupSync(callbackA: SyncCallback, callbackB: SyncCallback, warmupCount: number, suiteState: unknown, hooks: SyncHooks): void {
   for (let i = 0; i < warmupCount; i++) {
     runSyncWithHooks(callbackA, suiteState, hooks);
     runSyncWithHooks(callbackB, suiteState, hooks);
   }
-}
-
-function runSyncWithHooks(callback: SyncCallback, suiteState: unknown, hooks: ComparativeHooks): void {
-  const iterState = hooks.setupEach ? hooks.setupEach(suiteState) : undefined;
-  try {
-    callback(suiteState, iterState);
-  } finally {
-    if (hooks.teardownEach) hooks.teardownEach(suiteState, iterState);
-  }
-}
-
-function measureSync(callback: SyncCallback, suiteState: unknown, durations: number[], hooks: ComparativeHooks, allowedErrorRate: number): number {
-  let errorCount = 0;
-  const iterState = hooks.setupEach ? hooks.setupEach(suiteState) : undefined;
-  try {
-    const t0 = nowInMillis();
-    callback(suiteState, iterState);
-    const t1 = nowInMillis();
-    durations.push(t1 - t0);
-  } catch (e) {
-    if (allowedErrorRate === 0) throw e;
-    errorCount = 1;
-  } finally {
-    if (hooks.teardownEach) hooks.teardownEach(suiteState, iterState);
-  }
-  return errorCount;
 }
 
 /**
@@ -66,7 +30,7 @@ export function toBeFasterThan(callbackA: SyncCallback, callbackB: SyncCallback,
 
   const count = options.iterations;
   const confidence = options.confidence ?? 0.95;
-  const hooks: ComparativeHooks = options;
+  const hooks: SyncHooks = options;
   const allowedErrorRate = options.allowedErrorRate ?? 0;
   const suiteState = hooks.setup ? hooks.setup() : undefined;
 
@@ -94,47 +58,11 @@ export function toBeFasterThan(callbackA: SyncCallback, callbackB: SyncCallback,
   }
 }
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any -- expect.extend erases generics at runtime
-type AsyncCallback = (...args: any[]) => Promise<unknown>;
-
-interface AsyncComparativeHooks {
-  setup?: () => unknown;
-  teardown?: (suiteState: unknown) => void | Promise<void>;
-  setupEach?: (suiteState: unknown) => unknown;
-  teardownEach?: (suiteState: unknown, iterState: unknown) => void | Promise<void>;
-}
-
-async function warmupAsync(callbackA: AsyncCallback, callbackB: AsyncCallback, warmupCount: number, suiteState: unknown, hooks: AsyncComparativeHooks): Promise<void> {
+async function warmupAsync(callbackA: AsyncCallback, callbackB: AsyncCallback, warmupCount: number, suiteState: unknown, hooks: AsyncHooks): Promise<void> {
   for (let i = 0; i < warmupCount; i++) {
     await runAsyncWithHooks(callbackA, suiteState, hooks);
     await runAsyncWithHooks(callbackB, suiteState, hooks);
   }
-}
-
-async function runAsyncWithHooks(callback: AsyncCallback, suiteState: unknown, hooks: AsyncComparativeHooks): Promise<void> {
-  const iterState = hooks.setupEach ? await hooks.setupEach(suiteState) : undefined;
-  try {
-    await callback(suiteState, iterState);
-  } finally {
-    if (hooks.teardownEach) await hooks.teardownEach(suiteState, iterState);
-  }
-}
-
-async function measureAsync(callback: AsyncCallback, suiteState: unknown, durations: number[], hooks: AsyncComparativeHooks, allowedErrorRate: number): Promise<number> {
-  let errorCount = 0;
-  const iterState = hooks.setupEach ? await hooks.setupEach(suiteState) : undefined;
-  try {
-    const t0 = nowInMillis();
-    await callback(suiteState, iterState);
-    const t1 = nowInMillis();
-    durations.push(t1 - t0);
-  } catch (e) {
-    if (allowedErrorRate === 0) throw e;
-    errorCount = 1;
-  } finally {
-    if (hooks.teardownEach) await hooks.teardownEach(suiteState, iterState);
-  }
-  return errorCount;
 }
 
 /**
@@ -158,7 +86,7 @@ export async function toResolveFasterThan(promiseA: AsyncCallback, promiseB: Asy
 
   const count = options.iterations;
   const confidence = options.confidence ?? 0.95;
-  const hooks: AsyncComparativeHooks = options;
+  const hooks: AsyncHooks = options;
   const allowedErrorRate = options.allowedErrorRate ?? 0;
   const suiteState = hooks.setup ? await hooks.setup() : undefined;
